refactor some files in domain
