--- app/src/main/java/com/washingtondcsquad/tudee/data/localSource/entities/CategoryEntity.kt
@@ -1,16 +1,0 @@
-package com.washingtondcsquad.tudee.data.localSource.entities
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.washingtondcsquad.tudee.data.utils.DataBaseConstants
-
-
-@Entity(tableName = DataBaseConstants.ROOM_DATABASE_NAME)
-data class CategoryEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long,
-    @ColumnInfo(name = "Category Title") val title: String,
-    @ColumnInfo(name = "Category Image") val image: String,
-    @ColumnInfo(name = "Task Count") val taskCount: Int,
-
-    )
