+++ app/src/main/java/com/washingtondcsquad/tudee/data/localSource/utils/Extensions.kt
@@ -2,9 +2,7 @@ package com.washingtondcsquad.tudee.data.localSource.utils
 
 import java.security.MessageDigest
 
-
 fun String.toMD5(): String {
-    return MessageDigest.getInstance("MD5")
-        .digest(this.toByteArray())
+    return MessageDigest.getInstance("MD5").digest(this.toByteArray())
         .joinToString("") { "%02x".format(it) }
 }
