+++ app/src/main/java/com/washingtondcsquad/tudee/data/services/AppPreferencesServiceImpl.kt
@@ -1,48 +1,43 @@
 package com.washingtondcsquad.tudee.data.services
 
-import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.booleanPreferencesKey
 import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.preferencesDataStore
 import com.washingtondcsquad.tudee.domain.services.AppPreferencesService
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-
-val Context.appPreferencesDataStore: DataStore<Preferences> by preferencesDataStore(name = "app_preferences")
 
 class AppPreferencesServiceImpl(
     private val dataStore: DataStore<Preferences>
 ) : AppPreferencesService {
 
-    private object PreferencesKeys {
+    companion object PreferencesKeys {
         val HAS_ONBOARDING_BEEN_SHOWN = booleanPreferencesKey("has_onboarding_been_shown")
         val DARK_MODE_ENABLED = booleanPreferencesKey("dark_mode_enabled")
     }
 
     override fun hasOnboardingBeenShown(): Flow<Boolean> {
         return dataStore.data.map { preferences ->
-            preferences[PreferencesKeys.HAS_ONBOARDING_BEEN_SHOWN] ?: false
+            preferences[HAS_ONBOARDING_BEEN_SHOWN] == true
         }
     }
 
     override suspend fun setOnboardingShown() {
         dataStore.edit { preferences ->
-            preferences[PreferencesKeys.HAS_ONBOARDING_BEEN_SHOWN] = true
+            preferences[HAS_ONBOARDING_BEEN_SHOWN] = true
         }
     }
 
     override fun isDarkModeEnabled(): Flow<Boolean> {
         return dataStore.data.map { preferences ->
-            preferences[PreferencesKeys.DARK_MODE_ENABLED] ?: false
+            preferences[DARK_MODE_ENABLED] == true
         }
     }
 
-    override suspend fun setDarkModeEnabled(isEnabled: Boolean) {
+    override suspend fun setDarkMode(isEnabled: Boolean) {
         dataStore.edit { preferences ->
-            preferences[PreferencesKeys.DARK_MODE_ENABLED] = isEnabled
+            preferences[DARK_MODE_ENABLED] = isEnabled
         }
     }
-
 }
