--- app/src/main/java/com/washingtondcsquad/tudee/domain/services/LocaleService.kt
@@ -1,5 +1,0 @@
-package com.washingtondcsquad.tudee.domain.services
-
-interface LocaleService {
-    suspend fun switchAppTheme(isDarkTheme:Boolean)
-}
